New chemical: current()
The current() chemical allows you to look at the current element that
has been returned from next().

Also found a bug in the trait() function that needlessly returned a
closure rather than the class/function it was wrapping. Since it's only
job is to add the class/function to the `it` class, it didn't need to
also return the wrapper function.